+++ Tests/ContentfulPersistenceTests/ComplexTestModels/ComplexAsset.swift
@@ -1,0 +1,17 @@
+//
+//  ComplexAsset.swift
+//
+//
+//  Created by JP Wright on 31/03/16.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(ComplexAsset)
+class ComplexAsset: NSManagedObject {
+
+    // Insert code here to add functionality to your managed object subclass
+    
+}