+++ Tests/ContentfulPersistenceTests/ComplexTestModels/ComplexSyncInfo.swift
@@ -1,0 +1,18 @@
+//
+//  ComplexSyncInfo.swift
+//
+//
+//  Created by JP Wright on 31/03/16.
+//
+//
+
+import Foundation
+import CoreData
+import ContentfulPersistence
+
+@objc(ComplexSyncInfo)
+class ComplexSyncInfo: NSManagedObject {
+
+    // Insert code here to add functionality to your managed object subclass
+    
+}