+++ Tests/ContentfulPersistenceTests/ComplexTestModels/SingleRecord.swift
@@ -1,0 +1,15 @@
+//
+//  SingleRecord.swift
+//  ContentfulPersistence
+//
+//  Created by JP Wright on 12.07.17.
+//  Copyright © 2017 Contentful GmbH. All rights reserved.
+//
+
+import Foundation
+import CoreData
+
+@objc(SingleRecord)
+class SingleRecord: NSManagedObject {
+    // Insert code here to add functionality to your managed object subclass
+}