+++ Tests/ContentfulPersistenceTests/SynchronizationManagerTests.swift
@@ -7,15 +7,53 @@
 //
 
 @testable import ContentfulPersistence
+import ObjectMapper
 import Contentful
 import Nimble
 import Quick
+import CoreData
 
 typealias TestFunc = (() -> ()) throws -> ()
 
-class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
+class ContentfulPersistenceTests: QuickSpec {
+
     let assetPredicate = NSPredicate(format: "id == 'bXvdSYHB3Guy2uUmuEco8'")
     let postPredicate = NSPredicate(format: "id == '1asN98Ph3mUiCYIYiiqwko'")
+
+
+    let storeURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).last?.appendingPathComponent("Test.sqlite")
+
+    func deleteCoreDataStore() {
+        guard FileManager.default.fileExists(atPath: self.storeURL!.absoluteString) == true else { return }
+
+        try! FileManager.default.removeItem(at: self.storeURL!)
+        try! FileManager.default.removeItem(at: append("-shm", to: self.storeURL!))
+        try! FileManager.default.removeItem(at: append("-wal", to: self.storeURL!))
+    }
+
+    lazy var managedObjectContext: NSManagedObjectContext = {
+        let modelURL = Bundle(for: type(of: self)).url(forResource: "Test", withExtension: "momd")
+        let mom = NSManagedObjectModel(contentsOf: modelURL!)
+        expect(mom).toNot(beNil())
+
+        let psc = NSPersistentStoreCoordinator(managedObjectModel: mom!)
+
+        do {
+            var store = try psc.addPersistentStore(ofType: NSSQLiteStoreType, configurationName: nil, at: self.storeURL!, options: nil)
+            expect(store).toNot(beNil())
+        } catch {
+            XCTAssert(false, "Recreating the persistent store SQL files should not throw an error")
+        }
+
+        var managedObjectContext = NSManagedObjectContext(concurrencyType: .mainQueueConcurrencyType)
+        managedObjectContext.persistentStoreCoordinator = psc
+        return managedObjectContext
+    }()
+
+    func append(_ string: String, to fileURL: URL) -> URL {
+        let pathString = fileURL.path.appending(string)
+        return URL(fileURLWithPath: pathString)
+    }
 
     var client: Client!
 
@@ -30,13 +68,14 @@ class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
             self.client.initialSync() { result in
                 expect(result.value).toNot(beNil())
 
-                do {
-                    let posts: [Post] = try self.store.fetchAll(type: Post.self, predicate: NSPredicate(value: true))
-                    expect(posts.count).to(equal(2))
-
-                    try expectations(done)
-                } catch {
-                    XCTAssert(false, "Fetching posts should not throw an error")
+                self.managedObjectContext.perform {
+                    do {
+                        let posts: [Post] = try self.store.fetchAll(type: Post.self, predicate: NSPredicate(value: true))
+                        expect(posts.count).to(equal(2))
+                        try expectations(done)
+                    } catch {
+                        XCTAssert(false, "Fetching posts should not throw an error")
+                    }
                 }
             }
         }
@@ -57,12 +96,85 @@ class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
             self.deleteCoreDataStore()
         }
 
+        it("can correctly infer property mapping") {
+            // We must have a space first to pass in locale information.
+            let spaceMap = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("space"))
+            let space = try! Space(map: spaceMap)
+
+            let localesContext = space.localizationContext
+
+            let map = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("single-author"), context: localesContext)
+            let author = try! Entry(map: map)
+
+            let expectedPropertyMapping: [FieldName: String] = [
+                "name": "name",
+                "website": "website",
+                "biography": "biography"
+            ]
+            let authorPropertyMapping = self.sync.propertyMapping(for: Author.self, and: author.fields)
+            expect(authorPropertyMapping).to(equal(expectedPropertyMapping))
+        }
+
+        it("can correctly infer property mapping with explicitly defined mapping") {
+            // We must have a space first to pass in locale information.
+            let spaceMap = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("space"))
+            let space = try! Space(map: spaceMap)
+
+            let localesContext = space.localizationContext
+
+            let map = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("single-post"), context: localesContext)
+            let post = try! Entry(map: map)
+
+            let expectedPropertyMapping: [FieldName: String] = [
+                "title": "title"
+            ]
+            let postPropertyMapping = self.sync.propertyMapping(for: Post.self, and: post.fields)
+            expect(postPropertyMapping).to(equal(expectedPropertyMapping))
+        }
+
+        it("can correctly infer relationship mapping") {
+            // We must have a space first to pass in locale information.
+            let spaceMap = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("space"))
+            let space = try! Space(map: spaceMap)
+
+            let localesContext = space.localizationContext
+
+            let map = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("single-author"), context: localesContext)
+            let author = try! Entry(map: map)
+
+            let expectedRelationshipMapping: [FieldName: String] = [
+                "createdEntries": "createdEntries",
+                "profilePhoto": "profilePhoto"
+            ]
+            let authorRelationshipMapping = self.sync.relationshipMapping(for: Author.self, and: author.fields)
+            expect(authorRelationshipMapping).to(equal(expectedRelationshipMapping))
+        }
+
+        it("can correctly infer relationship mapping with explicitly defined mapping") {
+            // We must have a space first to pass in locale information.
+            let spaceMap = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("space"))
+            let space = try! Space(map: spaceMap)
+
+            let localesContext = space.localizationContext
+
+            let map = Map(mappingType: .fromJSON, JSON: TestHelpers.jsonData("single-post"), context: localesContext)
+            let post = try! Entry(map: map)
+
+            let expectedRelationshipMapping: [FieldName: String] = [
+                "featuredImage": "theFeaturedImage"
+            ]
+            let postRelationshiopMapping = self.sync.relationshipMapping(for: Post.self, and: post.fields)
+            expect(postRelationshiopMapping).to(equal(expectedRelationshipMapping))
+        }
+
         it("can store SyncTokens") { waitUntil(timeout: 10) { done in
             self.client.initialSync() { result in
-                expect(result.value!.assets.count).to(beGreaterThan(0))
-                expect(self.sync.syncToken?.characters.count).to(beGreaterThan(0))
+                self.managedObjectContext.perform {
+                    expect(result.value!.assets.count).to(beGreaterThan(0))
+                    expect(self.sync.syncToken?.characters.count).to(beGreaterThan(0))
 
-                done()
+                    done()
+                }
             }
         } }
 
@@ -77,25 +189,26 @@ class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
         it("can store Assets") { waitUntil(timeout: 10) { done in
             self.client.initialSync() { result in
 
-                do {
-                    let assets: [Asset] = try self.store.fetchAll(type: Asset.self, predicate: NSPredicate(value: true))
-                    expect(assets.count).to(equal(6))
+                self.managedObjectContext.perform {
+                    do {
+                        let assets: [Asset] = try self.store.fetchAll(type: Asset.self, predicate: NSPredicate(value: true))
+                        expect(assets.count).to(equal(6))
 
-                    let alice: Asset? = try self.store.fetchAll(type: Asset.self, predicate: self.assetPredicate).first
-                    expect(alice).toNot(beNil())
-                    expect(alice?.title).to(equal("Alice in Wonderland"))
-                    expect(alice?.urlString).to(equal("https://images.contentful.com/dqpnpm0n4e75/bXvdSYHB3Guy2uUmuEco8/608761ef6c0ef23815b410d5629208f9/alice-in-wonderland.gif"))
-                } catch {
-                    XCTAssert(false, "Fetching asset(s) should not throw an error")
+                        let alice: Asset? = try! self.store.fetchAll(type: Asset.self, predicate: self.assetPredicate).first
+                        expect(alice).toNot(beNil())
+                        expect(alice?.title).to(equal("Alice in Wonderland"))
+                        expect(alice?.urlString).to(equal("https://images.contentful.com/dqpnpm0n4e75/bXvdSYHB3Guy2uUmuEco8/608761ef6c0ef23815b410d5629208f9/alice-in-wonderland.gif"))
+                    } catch {
+                        XCTAssert(false, "Fetching asset(s) should not throw an error")
+                    }
+                    done()
                 }
-
-                done()
             }
         } }
 
         it("can store Entries") {
             self.postTests { done in
-                let post: Post? = try self.store.fetchAll(type: Post.self, predicate: self.postPredicate).first
+                let post: Post? = try! self.store.fetchAll(type: Post.self, predicate: self.postPredicate).first
                 expect(post).toNot(beNil())
                 expect(post?.title).to(equal("Down the Rabbit Hole"))
                 done()
@@ -107,9 +220,9 @@ class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
                 let post: Post? = try self.store.fetchAll(type: Post.self, predicate: self.postPredicate).first
                 expect(post).toNot(beNil())
 
-                expect(post?.featuredImage).toNot(beNil())
-                expect(post?.featuredImage?.urlString).toNot(beNil())
-                expect(post?.featuredImage?.urlString).to(equal("https://images.contentful.com/dqpnpm0n4e75/bXvdSYHB3Guy2uUmuEco8/608761ef6c0ef23815b410d5629208f9/alice-in-wonderland.gif"))
+                expect(post?.theFeaturedImage).toNot(beNil())
+                expect(post?.theFeaturedImage?.urlString).toNot(beNil())
+                expect(post?.theFeaturedImage?.urlString).to(equal("https://images.contentful.com/dqpnpm0n4e75/bXvdSYHB3Guy2uUmuEco8/608761ef6c0ef23815b410d5629208f9/alice-in-wonderland.gif"))
 
                 done()
             }
@@ -139,6 +252,7 @@ class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
                 expect(post).toNot(beNil())
                 expect(post?.comments).to(beNil())
                 expect(post?.title).toNot(beNil())
+                expect(post?.theFeaturedImage).toNot(beNil())
                 done()
             }
         }
@@ -155,13 +269,13 @@ class ContentfulPersistenceTests: ContentfulPersistenceTestBase {
             }
         }
 
-        it("can determine relationships of a type") {
+        it("can determine relationships of a CoreData type") {
             let store = CoreDataStore(context: self.managedObjectContext)
 
             do {
                 let relationships = try store.relationships(for: Post.self)
-
-                expect(relationships).to(equal(["author", "category", "featuredImage"]))
+                let expectedRelationships = ["author", "category", "theFeaturedImage"]
+                expect(relationships).to(equal(expectedRelationships))
             } catch {
                 XCTAssert(false, "Storing relationships for Posts should not throw an error")
             }